Create and import modules for the add and remove functions
